requirements.txt
awscli
boto3
environs==6.0.0
PyYAML
